third_party/WebKit/LayoutTests/external/wpt/html/webappapis/scripting/events/event-handler-processing-algorithm-error/window-synthetic-event-expected.txt

This is a testharness.js-based test.
FAIL error event is normal (return true does not cancel; one arg) on Window, with a synthetic Event assert_equals: expected false but got true
Harness: the test ran to completion.
